feat: refactor mutations
